Moving Image_Asset classews into position
