+++ fluid/CMakeLists.txt
@@ -114,7 +114,6 @@ set(CPPFILES
   fluid_icon.cxx
   app/args.cxx
   app/Snap_Action.cxx
-  app/Image_Asset.cxx
   app/history.cxx
   app/Menu.cxx
   app/shell_command.cxx
@@ -144,6 +143,8 @@ set(CPPFILES
   panels/widget_panel/Grid_Child_Tab.cxx
   proj/align_widget.cxx
   proj/i18n.cxx
+  proj/Image_Asset.cxx
+  proj/Image_Asset_Map.cxx
   proj/mergeback.cxx
   proj/undo.cxx
   rsrcs/pixmaps.cxx
@@ -168,7 +169,6 @@ set(HEADERFILES
   fluid_icon.h
   app/args.h
   app/Snap_Action.h
-  app/Image_Asset.h
   app/history.h
   app/Menu.h
   app/shell_command.h
@@ -197,6 +197,8 @@ set(HEADERFILES
   panels/widget_panel/Grid_Tab.h
   panels/widget_panel/Grid_Child_Tab.h
   proj/align_widget.h
+  proj/Image_Asset.h
+  proj/Image_Asset_Map.h
   proj/mergeback.h
   proj/i18n.h
   proj/undo.h
