+++ fluid/nodes/Function_Node.h
@@ -19,7 +19,7 @@
 
 #include "nodes/Node.h"
 
-#include "app/Image_Asset.h"
+#include "proj/Image_Asset.h"
 #ifdef _WIN32
 #include "tools/ExternalCodeEditor_WIN32.h"
 #else
